+++ packages/client/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Fixed
 
-- Error when unable to track a transaction because we havne't fully recovered from a `stop` event
+- Error when unable to track a transaction because we haven't fully recovered from a `stop` event
 
 ### Added